Small macros to make moving metrics easy.
